Sorta works, but boolean indexing slow This is a kludgy, only somewhat broken way to make references to the current document work. It passes all but one of the tests in my test suite.
It passes the tokens representing an indexer into
the parse_index function every time indexing is
done and rebuilds the AST using the document that
was passed into that level of the indexer.

This approach is clearly wasteful, because the values
of most things can and should be precomputed so that
all you have is a function that performs predefined
computations on a variable document passed into it.